Design For Assembly In KiCad Project Files...
This repo contains all the KiCad project files, CSV files and gerbers generating during the creation of the "Design for Assembly" post at http://rheingoldheavy.com/design-assembly-kicad.  It is designed to walk you through creating a circuit board with the essential elements in place that would allow you to use a contract manufacturer to assemble your board for you.